+++ hedera-node/hedera-token-service-impl/src/test/java/com/hedera/node/app/service/token/impl/test/schemas/V0490TokenSchemaTest.java
@@ -21,14 +21,6 @@ import static com.hedera.node.app.service.token.impl.schemas.V0490TokenSchema.AC
 import static com.hedera.node.app.service.token.impl.schemas.V0490TokenSchema.ALIASES_KEY;
 import static com.hedera.node.app.service.token.impl.schemas.V0490TokenSchema.STAKING_INFO_KEY;
 import static com.hedera.node.app.service.token.impl.schemas.V0490TokenSchema.STAKING_NETWORK_REWARDS_KEY;
-import static com.hedera.node.app.service.token.impl.test.handlers.staking.EndOfStakingPeriodUpdaterTest.NODE_NUM_1;
-import static com.hedera.node.app.service.token.impl.test.handlers.staking.EndOfStakingPeriodUpdaterTest.NODE_NUM_2;
-import static com.hedera.node.app.service.token.impl.test.handlers.staking.EndOfStakingPeriodUpdaterTest.NODE_NUM_3;
-import static com.hedera.node.app.service.token.impl.test.handlers.staking.EndOfStakingPeriodUpdaterTest.NODE_NUM_4;
-import static com.hedera.node.app.service.token.impl.test.handlers.staking.EndOfStakingPeriodUpdaterTest.NODE_NUM_8;
-import static com.hedera.node.app.service.token.impl.test.handlers.staking.EndOfStakingPeriodUpdaterTest.STAKING_INFO_1;
-import static com.hedera.node.app.service.token.impl.test.handlers.staking.EndOfStakingPeriodUpdaterTest.STAKING_INFO_2;
-import static com.hedera.node.app.service.token.impl.test.handlers.staking.EndOfStakingPeriodUpdaterTest.STAKING_INFO_3;
 import static com.hedera.node.app.service.token.impl.test.schemas.SyntheticAccountsData.DEFAULT_NUM_SYSTEM_ACCOUNTS;
 import static com.hedera.node.app.service.token.impl.test.schemas.SyntheticAccountsData.EVM_ADDRESSES;
 import static com.hedera.node.app.service.token.impl.test.schemas.SyntheticAccountsData.NUM_RESERVED_SYSTEM_ENTITIES;
@@ -258,52 +250,6 @@ final class V0490TokenSchemaTest {
         }
     }
 
-    @Test
-    void marksNonExistingNodesToDeletedInStateAndAddsNewNodesToState() {
-        accounts = MapWritableKVState.<AccountID, Account>builder(V0490TokenSchema.ACCOUNTS_KEY)
-                .build();
-        // State has nodeIds 1, 2, 3
-        final var stakingInfosState = new MapWritableKVState.Builder<EntityNumber, StakingNodeInfo>(STAKING_INFO_KEY)
-                .value(NODE_NUM_1, STAKING_INFO_1)
-                .value(NODE_NUM_2, STAKING_INFO_2)
-                .value(NODE_NUM_3, STAKING_INFO_3)
-                .build();
-        final var previousStates = newStatesInstance(
-                accounts,
-                MapWritableKVState.<Bytes, AccountID>builder(ALIASES_KEY).build(),
-                newWritableEntityIdState(),
-                stakingInfosState);
-        newStates = newStatesInstance(
-                accounts,
-                MapWritableKVState.<Bytes, AccountID>builder(ALIASES_KEY).build(),
-                newWritableEntityIdState(),
-                stakingInfosState);
-        entityIdStore = new WritableEntityIdStore(newStates);
-        // Platform address book has node Ids 2, 4, 8
-        networkInfo = new FakeNetworkInfo();
-        config = buildConfig(DEFAULT_NUM_SYSTEM_ACCOUNTS, true);
-
-        final var schema = newSubjectWithAllExpected();
-        // When we call restart, the state will be updated to mark node 1 and 3 as deleted
-        schema.restart(new MigrationContextImpl(
-                previousStates, newStates, config, networkInfo, entityIdStore, null, new HashMap<>()));
-        final var updatedStates = newStates.get(STAKING_INFO_KEY);
-        // marks nodes 1, 2 as deleted
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_1)).deleted()).isTrue();
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_2)).deleted()).isFalse();
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_3)).deleted()).isTrue();
-        // Also adds node 4 to the state
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_4)).deleted()).isFalse();
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_4)).weight()).isZero();
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_4)).minStake()).isZero();
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_4)).maxStake()).isEqualTo(1666666666666666666L);
-        // Also adds node 8 to the state
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_8)).deleted()).isFalse();
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_8)).weight()).isZero();
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_8)).minStake()).isZero();
-        assertThat(((StakingNodeInfo) updatedStates.get(NODE_NUM_8)).maxStake()).isEqualTo(1666666666666666666L);
-    }
-
     private WritableSingletonState<EntityNumber> newWritableEntityIdState() {
         return new WritableSingletonStateBase<>(
                 V0490EntityIdSchema.ENTITY_ID_STATE_KEY, () -> new EntityNumber(BEGINNING_ENTITY_ID), c -> {});
@@ -319,21 +265,6 @@ final class V0490TokenSchemaTest {
                 .state(aliases)
                 .state(MapWritableKVState.builder(V0490TokenSchema.STAKING_INFO_KEY)
                         .build())
-                .state(new WritableSingletonStateBase<>(STAKING_NETWORK_REWARDS_KEY, () -> null, c -> {}))
-                .state(entityIdState)
-                .build();
-    }
-
-    private MapWritableStates newStatesInstance(
-            final MapWritableKVState<AccountID, Account> accts,
-            final MapWritableKVState<Bytes, AccountID> aliases,
-            final WritableSingletonState<EntityNumber> entityIdState,
-            final MapWritableKVState<EntityNumber, StakingNodeInfo> stakingInfo) {
-        //noinspection ReturnOfNull
-        return MapWritableStates.builder()
-                .state(accts)
-                .state(aliases)
-                .state(stakingInfo)
                 .state(new WritableSingletonStateBase<>(STAKING_NETWORK_REWARDS_KEY, () -> null, c -> {}))
                 .state(entityIdState)
                 .build();
